feat: improve AC-04.01 evaluation coverage (#158)
* feat: improve AC-04.01 evalaution coverage

This change adds logic to check the value of `can_approve_pull_request_reviews`
as the creation and approval of PRs is arguably a write permission grant.

See https://docs.github.com/en/rest/actions/permissions?apiVersion=2022-11-28#get-default-workflow-permissions-for-a-repository and https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/enabling-features-for-your-repository/managing-github-actions-settings-for-a-repository#preventing-github-actions-from-creating-or-approving-pull-requests for reference to behavior
of the GitHub configuration and REST API.

Signed-off-by: Travis Truman <trumant@gmail.com>

* fix: validate workflows enabled before default workflow perms

This change fixes a subtle bug in the implementation of AC-04.01
where default workflow permissions can be configured to safe/recommended
values, but if GHA is not enabled, the configuration is rendered moot
with respect to the project's CI/CD pipeline least privilege implementation.

In scenarios where GHA is disabled altogether, the evaluation of this
control requirement now returns `Needs Review`

Signed-off-by: Travis Truman <trumant@gmail.com>

---------

Signed-off-by: Travis Truman <trumant@gmail.com>